requirements.txt
flask
yt-dlp
openai-whisper
requests
gunicorn 